#7710 move out file operation error and extended from yustException not exception
